Generic discover model clean-up
